[FIX] mass_mailing_sms: improve twilio results support when testing
'success' or 'sent' can be received when succeeding, leading to
a more correct notification message. Also fix a double failure
reason appearing in notification message.

Task-5053537
